+++ app/Http/QueryFilters/BillingProfileFilters.php
@@ -13,11 +13,18 @@ class BillingProfileFilters extends QueryFilter
 
     public function owner(string $search): Builder
     {
-        return $this->builder->where('billing_profiles.owner_id', $search);
+        return $this->builder->whereHas('owner', function ($query) use ($search) {
+            return $query->where('name', 'ILIKE', "%$search%");
+        });
     }
 
     public function state(string $search): Builder
     {
         return $this->builder->where('state', $search);
+    }
+
+    public function legal_person(bool $search): Builder
+    {
+        return $this->builder->where('legal_person', $search);
     }
 }