+++ config/services.php
@@ -50,6 +50,10 @@ return [
         'secret' => env('FIC_SECRET'),
     ],
 
+    'maps' => [
+        'key' => env('GOOGLE_MAPS_API_KEY'),
+    ],
+
     'google' => [
         'client_id' => env('GOOGLE_CLIENT_ID'),
         'client_secret' => env('GOOGLE_CLIENT_SECRET'),