+++ resources/views/emails/billing-profile/approved.blade.php
@@ -1,7 +1,7 @@
 @component('mail::message')
 # Società riconosciuta!
 
-Congratulazioni, la società **{{ $billingProfile->business_name }}** è stata
+Congratulazioni, la società **{{ $billingProfile->name }}** è stata
 riconosciuta dal nostro staff! Ora sarai in grado di visualizzare **prezzi** e
 **disponibilità** sul nostro portale.