+++ resources/views/layouts/app.blade.php
@@ -10,7 +10,7 @@
     <title>{{ config('app.name', 'The BeerWay') }}</title>
 
     <!-- Scripts -->
-    <script src="https://maps.googleapis.com/maps/api/js?key={{ env('GOOGLE_MAPS_API_KEY') }}&libraries=places"></script>
+    <script src="https://maps.googleapis.com/maps/api/js?key={{ config('services.maps.key') }}&libraries=places"></script>
     <script src="{{ asset('js/app.js') }}" defer></script>
 
     <!-- Fonts -->